test(TrackedTripUtils): Extract end tracking code.
